project/requirements.txt
Django==4.1.3 
python-dotenv==1.0.0 
langchain==0.0.100 
google-generativeai==0.2.1
langchain-google-genai==0.1.0 
langchain-core==0.0.100 
